requirements.txt
accelerate==0.31.0
bitsandbytes==0.45.5
datasets<=3.5.1
evaluate==0.4.3
huggingface-hub==0.30.2
langchain==0.3.25
matplotlib==3.9.3
openpyxl==3.1.5
pandas==2.2.3
peft==0.13.0
proto-plus==1.26.1
protobuf==4.25.7
pydantic==2.10.6
python-dotenv==1.0.1
regex==2024.5.15
scikit-learn>=1.5.0
torch==2.6.0+cu124
torchvision==0.21.0+cu124
tqdm==4.66.5
transformers==4.41.2
trio==0.29.0
trio-websocket==0.12.2
triton==3.2.0
trl==0.11.0
uvicorn==0.34.0
wandb==0.19.10
yacs==0.1.8
--extra-index-url https://download.pytorch.org/whl/cu124